+++ src/lib/output/themes/MarkedPlugin.tsx
@@ -1,13 +1,24 @@
 import * as fs from "fs";
 import * as Path from "path";
-import * as Marked from "marked";
+import markdown from "markdown-it";
 
 import { Component, ContextAwareRendererComponent } from "../components";
 import { type RendererEvent, MarkdownEvent, type PageEvent } from "../events";
-import { Option, readFile, copySync, isFile, JSX, renderElement } from "../../utils";
+import { Option, readFile, copySync, isFile, type Logger } from "../../utils";
 import { highlight, isSupportedLanguage } from "../../utils/highlighter";
 import type { Theme } from "shiki";
 import { escapeHtml, getTextContent } from "../../utils/html";
+import type { DefaultTheme } from "..";
+import { Slugger } from "./default/DefaultTheme";
+
+let defaultSlugger: Slugger | undefined;
+function getDefaultSlugger(logger: Logger) {
+    if (!defaultSlugger) {
+        logger.warn(logger.i18n.custom_theme_does_not_define_getSlugger());
+        defaultSlugger = new Slugger();
+    }
+    return defaultSlugger;
+}
 
 /**
  * Implements markdown and relativeURL helpers for templates.
@@ -26,6 +37,8 @@ export class MarkedPlugin extends ContextAwareRendererComponent {
 
     @Option("darkHighlightTheme")
     accessor darkTheme!: Theme;
+
+    private parser?: MarkdownIt;
 
     /**
      * The path referenced files are located in.
@@ -127,7 +140,7 @@ export class MarkedPlugin extends ContextAwareRendererComponent {
     protected override onBeginRenderer(event: RendererEvent) {
         super.onBeginRenderer(event);
 
-        Marked.marked.setOptions(this.createMarkedOptions());
+        this.setupParser();
 
         delete this.includes;
         if (this.includeSource) {
@@ -151,56 +164,66 @@ export class MarkedPlugin extends ContextAwareRendererComponent {
         }
     }
 
+    private getSlugger() {
+        if ("getSlugger" in this.owner.theme!) {
+            return (this.owner.theme as DefaultTheme).getSlugger(this.page!.model);
+        }
+        return getDefaultSlugger(this.application.logger);
+    }
+
     /**
      * Creates an object with options that are passed to the markdown parser.
      *
      * @returns The options object for the markdown parser.
      */
-    private createMarkedOptions(): Marked.marked.MarkedOptions {
-        const markedOptions = (this.application.options.getValue("markedOptions") ?? {}) as Marked.marked.MarkedOptions;
+    private setupParser() {
+        this.parser = markdown({
+            ...(this.application.options.getValue("markdownItOptions") as {}),
+            highlight: (code, lang) => {
+                code = highlight(code, lang || "ts");
+                code = code.replace(/\n$/, "") + "\n";
 
-        // Set some default values if they are not specified via the TypeDoc option
-        markedOptions.highlight ??= (text, lang) => this.getHighlighted(text, lang);
+                if (!lang) {
+                    return `<pre><code>${code}</code><button>Copy</button></pre>\n`;
+                }
 
-        if (!markedOptions.renderer) {
-            markedOptions.renderer = new Marked.Renderer();
+                return `<pre><code class="${escapeHtml(lang)}">${code}</code><button>Copy</button></pre>\n`;
+            },
+        });
 
-            markedOptions.renderer.link = (href, title, text) => {
-                // Prefix the #anchor links `#md:`.
-                const target = href?.replace(/^#(?:md:)?(.+)/, "#md:$1") || undefined;
-                return renderElement(
-                    <a href={target} title={title || undefined}>
-                        <JSX.Raw html={text} />
-                    </a>,
-                );
-            };
+        const loader = this.application.options.getValue("markdownItLoader");
+        loader(this.parser);
 
-            markedOptions.renderer.heading = (text, level, _, slugger) => {
-                const slug = slugger.slug(text);
-                // Prefix the slug with an extra `md:` to prevent conflicts with TypeDoc's anchors.
-                this.page!.pageHeadings.push({
-                    link: `#md:${slug}`,
-                    text: getTextContent(text),
-                    level,
-                });
-                const H = `h${level}`;
-                return renderElement(
-                    <>
-                        <a id={`md:${slug}`} class="tsd-anchor" />
-                        <H>
-                            <a href={`#md:${slug}`}>
-                                <JSX.Raw html={text} />
-                            </a>
-                        </H>
-                    </>,
-                );
-            };
-            markedOptions.renderer.code = renderCode;
-        }
+        // Add anchor links for headings in readme, and add them to the "On this page" section
+        this.parser.renderer.rules["heading_open"] = (tokens, idx) => {
+            const token = tokens[idx];
+            const content = tokens[idx + 1].content;
+            const level = token.markup.length;
 
-        markedOptions.mangle ??= false; // See https://github.com/TypeStrong/typedoc/issues/1395
+            const slug = this.getSlugger().slug(content);
+            // Prefix the slug with an extra `md:` to prevent conflicts with TypeDoc's anchors.
+            this.page!.pageHeadings.push({
+                link: `#md:${slug}`,
+                text: getTextContent(content),
+                level,
+            });
 
-        return markedOptions;
+            return `<a id="md:${slug}" class="tsd-anchor"></a><${token.tag}><a href="#md:${slug}">`;
+        };
+        this.parser.renderer.rules["heading_close"] = (tokens, idx) => {
+            return `</a></${tokens[idx].tag}>`;
+        };
+
+        // Rewrite anchor links inline in a readme file to links targeting the `md:` prefixed anchors
+        // that TypeDoc creates.
+        this.parser.renderer.rules["link_open"] = (tokens, idx, options, _env, self) => {
+            const token = tokens[idx];
+            const href = token.attrGet("href")?.replace(/^#(?:md:)?(.+)/, "#md:$1");
+            if (href) {
+                token.attrSet("href", href);
+            }
+            return self.renderToken(tokens, idx, options);
+        };
     }
 
     /**
@@ -209,29 +232,6 @@ export class MarkedPlugin extends ContextAwareRendererComponent {
      * @param event
      */
     onParseMarkdown(event: MarkdownEvent) {
-        event.parsedText = Marked.marked(event.parsedText);
+        event.parsedText = this.parser!.render(event.parsedText);
     }
-}
-
-// Basically a copy/paste of Marked's code, with the addition of the button
-// https://github.com/markedjs/marked/blob/v4.3.0/src/Renderer.js#L15-L39
-function renderCode(this: Marked.marked.Renderer, code: string, infostring: string | undefined, escaped: boolean) {
-    const lang = (infostring || "").match(/\S*/)![0];
-    if (this.options.highlight) {
-        const out = this.options.highlight(code, lang);
-        if (out != null && out !== code) {
-            escaped = true;
-            code = out;
-        }
-    }
-
-    code = code.replace(/\n$/, "") + "\n";
-
-    if (!lang) {
-        return `<pre><code>${escaped ? code : escapeHtml(code)}</code><button>Copy</button></pre>\n`;
-    }
-
-    return `<pre><code class="${this.options.langPrefix + escapeHtml(lang)}">${
-        escaped ? code : escapeHtml(code)
-    }</code><button>Copy</button></pre>\n`;
 }