+++ src/lib/utils/options/declaration.ts
@@ -134,7 +134,25 @@ export interface TypeDocOptionMap {
     lightHighlightTheme: ShikiTheme;
     darkHighlightTheme: ShikiTheme;
     customCss: string;
-    markedOptions: unknown;
+    markdownItOptions: unknown;
+    /**
+     * Will be called when TypeDoc is setting up the markdown parser to use to render markdown.
+     * Can be used to add markdown-it plugins to the parser with code like this:
+     *
+     * ```ts
+     * // typedoc.config.mjs
+     * import iterator from "markdown-it-for-inline";
+     * export default {
+     *     /** @type {MarkdownIt} *\/
+     *     markdownItLoader(parser) {
+     *         parser.use(iterator, "foo_replace", "text", function(tokens, idx) {
+     *             tokens[idx].content = tokens[idx].content.replace(/foo/g, 'bar');
+     *         });
+     *     }
+     * }
+     * ```
+     */
+    markdownItLoader: ManuallyValidatedOption<(parser: MarkdownIt) => void>;
     basePath: string;
     cname: string;
     githubPages: boolean;